+++ fv3core/pace/fv3core/initialization/__init__.py
@@ -1,4 +1,1 @@
-# from .baroclinic import init_baroclinic_state
-# from .tropical_cyclone import init_tc_state
-from .test_cases.initialize_baroclinic import init_baroclinic_state
-from .test_cases.initialize_tc import init_tc_state
+from .analytic_init.py import init_analytic_state
